requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.1
fastapi==0.123.9
h11==0.16.0
idna==3.11
kombu==5.6.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
redis==7.0.1
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
